[REFACTOR] utils 내부에 있는 css 파일 styles 내부로 이동
